Split type_ref
